small binary patch to fix crashing caused by the story skip scripts
AF3DN.P is an extra module added to the Steam version that handles the new rendering options, Steam integrations, etc. - It tracks your progress using a pointer which, for reasons, is set when you watch certain movie files. Since i changed the intro movie, it was no longer setting this pointer at the start of a new game (though it would figure it out if you saved & reloaded), so when it tried to check your achievement progress or whatever it would crash out. The patch changes the filename it wants from "disc00_30h.avi" to "disc00_01h.avi". If you turn off the story skip option it patches back to what it was before.